Selenium/requirements.txt
pyyaml
pytest
pytest-selenium
allure-pytest